+++ axx_reverse_charge/models/purchase/__init__.py
@@ -1,1 +1,2 @@
+from . import purchase_order
 from . import purchase_order_line
