+++ axx_reverse_charge/models/sale/__init__.py
@@ -1,1 +1,2 @@
+from . import sale_order
 from . import sale_order_line
