+++ axx_reverse_charge/views/product/product_category_views.xml
@@ -13,4 +13,16 @@
         </field>
     </record>
 
+    <record id="view_axx_product_category_property_form" model="ir.ui.view">
+        <field name="name">axx_product_category_property_form</field>
+        <field name="model">product.category</field>
+        <field name="inherit_id" ref="account.view_category_property_form"/>
+        <field name="arch" type="xml">
+            <xpath expr="//group/group/field[@name='property_account_expense_categ_id']" position="after">
+                <field name="axx_rc_income_acc_id" attrs="{'invisible': [('axx_is_rc_relevant', '=', False), ('axx_is_additional_service', '=', False)]}"/>
+                <field name="axx_rc_expense_acc_id" attrs="{'invisible': [('axx_is_rc_relevant', '=', False), ('axx_is_additional_service', '=', False)]}"/>
+            </xpath>
+        </field>
+    </record>
+
 </odoo>
